Added polysquare_add_test_main to add a main() to tests.
This allows the user to specify one custom library which provides
the main () function for the tests, for overriding the global test
environment and reporters.

A single library that provides main () can be linked in by
providing as the single argument to the TEST_MAIN option in
polysquare_add_test.